+++ CHANGELOG.md
@@ -21,3 +21,4 @@
  - add aliases `and`, `or`, `andFiltering` and `orFiltering`
  - force old style queries still using aliases
  - apply operator more times using arrays
+ - store queries with name